requirements.txt
Flask==2.3.3
werkzeug==2.3.6
python-dotenv==1.0.0
assemblyai==1.0.0
Gunicorn
